requirements.txt
python-dotenv==1.2.1
openai==2.21.0
openpyxl==3.1.5
pandas==3.0.1
sympy==1.14.0
scikit-learn==1.8.0
numpy==2.4.2
sentence-transformers==5.2.3
tqdm==4.67.3
langchain-openai>=0.2.0
langchain-core>=0.3.0
pydantic>=2.0.0
